feat: 프로젝트 디렉토리 변경 + facade 제거
