+++ src/main/java/com/hhplus/clean_architecture/domain/entity/base/TimeBaseEntity.java
@@ -1,0 +1,20 @@
+package com.hhplus.clean_architecture.domain.entity.base;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.EntityListeners;
+import jakarta.persistence.MappedSuperclass;
+import org.hibernate.annotations.ColumnDefault;
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
+import java.time.LocalDateTime;
+
+@EntityListeners(AuditingEntityListener.class)
+@MappedSuperclass
+public class TimeBaseEntity {
+
+    @CreatedDate
+    @Column(nullable = false, updatable = false)
+    @ColumnDefault("CURRENT_TIMESTAMP")
+    protected LocalDateTime createdAt;
+}
