+++ src/main/java/com/hhplus/clean_architecture/domain/exception/AlreadyEnrollmentException.java
@@ -1,0 +1,15 @@
+package com.hhplus.clean_architecture.domain.exception;
+
+public class AlreadyEnrollmentException extends RuntimeException {
+
+    private final String errorString;
+
+    public AlreadyEnrollmentException(String errorString, String message) {
+        super(message);
+        this.errorString = errorString;
+    }
+
+    public String getErrorString() {
+        return errorString;
+    }
+}
