+++ src/main/java/com/hhplus/clean_architecture/domain/repository/LectureRepository.java
@@ -1,0 +1,10 @@
+package com.hhplus.clean_architecture.domain.repository;
+
+import com.hhplus.clean_architecture.domain.entity.Lecture;
+
+import java.util.List;
+
+public interface LectureRepository {
+
+    List<Lecture> findAll();
+}
