--- src/main/java/com/hhplus/clean_architecture/dto/response/LectureTimeListResponse.java
@@ -1,19 +1,0 @@
-package com.hhplus.clean_architecture.dto.response;
-
-import lombok.Builder;
-import lombok.Getter;
-import lombok.Setter;
-
-import java.time.LocalDateTime;
-
-@Getter
-@Setter
-@Builder
-public class LectureTimeListResponse {
-
-    private String lectureId;
-
-    private int capacity;
-
-    private LocalDateTime lectureTime;
-}
