--- src/main/java/com/hhplus/clean_architecture/exception/LectureFullException.java
@@ -1,15 +1,0 @@
-package com.hhplus.clean_architecture.exception;
-
-public class LectureFullException extends RuntimeException {
-
-    private final String errorString;
-
-    public LectureFullException(String message) {
-        super(message);
-        this.errorString = "E500";
-    }
-
-    public String getErrorString() {
-        return errorString;
-    }
-}
