+++ src/main/java/com/hhplus/clean_architecture/infrastructure/persistence/LectureJpaRepository.java
@@ -1,0 +1,7 @@
+package com.hhplus.clean_architecture.infrastructure.persistence;
+
+import com.hhplus.clean_architecture.domain.entity.Lecture;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface LectureJpaRepository extends JpaRepository<Lecture, String> {
+}
