+++ src/main/java/com/hhplus/clean_architecture/interfaces/dto/response/LectureTimeListResponse.java
@@ -1,0 +1,11 @@
+package com.hhplus.clean_architecture.interfaces.dto.response;
+
+import java.time.LocalDateTime;
+
+public record LectureTimeListResponse(
+        String lectureId,
+        int capacity,
+        LocalDateTime lectureTime
+) {
+
+}
