--- src/main/java/com/hhplus/clean_architecture/repository/EnrollmentRepository.java
@@ -1,15 +1,0 @@
-package com.hhplus.clean_architecture.repository;
-
-import com.hhplus.clean_architecture.entity.Enrollment;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-
-public interface EnrollmentRepository extends JpaRepository<Enrollment, Long> {
-
-    List<Enrollment> findByUserId(String userId);
-
-    List<Enrollment> findAllByLectureTimeId(Long lectureTimeId);
-
-    long countByLectureTimeId(long lectureTimeId);
-}
