--- src/main/java/com/hhplus/clean_architecture/repository/LectureRepository.java
@@ -1,7 +1,0 @@
-package com.hhplus.clean_architecture.repository;
-
-import com.hhplus.clean_architecture.entity.Lecture;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface LectureRepository extends JpaRepository<Lecture, String> {
-}
