--- src/main/java/com/hhplus/clean_architecture/repository/UserRepository.java
@@ -1,7 +1,0 @@
-package com.hhplus.clean_architecture.repository;
-
-import com.hhplus.clean_architecture.entity.Users;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository<Users, String> {
-}
